+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - "Use RText Server" settings option (default: enabled). When it is enabled the `bake-rtext-service` is used to get the diagnostic information (can be found in Problems tab). Bake version 2.56.0 or higher is required.
+ - Initial support of multi-root workspace
 
 ## [0.8.0] - 2019-12-23
 ### Added